tests/wpt/tests/css/css-position/overlay/overlay-transition-property.html

<!DOCTYPE html>
<title>CSS Positioned Layout Module Test: 'overlay' is a valid transition-property value</title>
<link rel="help" href="https://drafts.csswg.org/css-position-4/#overlay">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<div id="target"></div>
<script>
  test_valid_value("transition-property", "overlay");
  test_valid_value("transition", "overlay 2s");

  test_computed_value("transition-property", "overlay");
  test_computed_value("transition", "overlay 1s");
</script>
